requirements.txt
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.25.0
vllm>=0.6.0
Pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
openai>=1.0.0
dashscope>=1.14.0
qwen-vl-utils>=0.0.8
t2v-metrics>=1.0.0
requests>=2.31.0
tabulate>=0.9.0
jsonlines>=4.0.0
prettytable>=3.9.0
